Create Keypad_Interfacing.md
While working with most electronic projects, the keypad is a common component and for a 4×4 keypad with
16 keys and it requires 8 digital I/O lines of the microcontroller device while interfacing. GPIO pins are valuable
assets and we need to reserve them as much as we can. This is an implementation of interfacing the 4x4 keypad using a
single analogue pin which enables one to save a considerable amount of the microcontroller pins for other uses.